Move everything to multiplayerController, Inputfield testing
